Implemented constraints parameter. You can now pass a constraints option to any of the media slots. The constraints option is an array. Examples:
"constraints" => array("width" => 300, "height" => 200) 
"constraints" => array("minimum-width" => 300, "minimum-height" => 200) 
"constraints" => array("aspect-width" => 4, "aspect-height" => 3) 

It makes sense to combine aspect-width and aspect-height with
minimum-width and minimum-height. 

TODO, I GUESS: I'm not checking the results that come back from the media
plugin to make sure they aren't spoofed with regard to the image
dimensions. That would be a weird thing for a person with editing
privileges to do!



git-svn-id: http://svn.symfony-project.com/plugins/pkMediaCMSSlotsPlugin/trunk@17061 ee427ae8-e902-0410-961c-c3ed070cd9f9